# contracts/auth-openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ExhibitOS reference authentication API",
    "version": "0.1.0",
    "description": "Tenant-bound reference password/session adapter and metadata policy, plus full public OES rights-validated private attachment delivery (permissions.download) and export eligibility (permissions.export). Legacy flat rights are rejected, without implicit permission migration. No anonymous publication, signed object URLs or OEX byte export. Configure one exact origin. Reference HTTP listener binds only to loopback in both modes; network mode requires an external TLS terminator."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local loopback example; must match configured origin."
    },
    {
      "url": "https://api.example.invalid",
      "description": "Synthetic network placeholder; replace with the exact configured HTTPS origin."
    }
  ],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "PostLoginToTenant",
        "summary": "Login to tenant",
        "description": "Requires existing credentials and tenant membership. Issues a cookie; use GET /auth/session to obtain session information and CSRF token. Login does not require a pre-existing CSRF token.",
        "security": [],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Must equal AUTH_ORIGIN exactly when supplied. Required on login and authenticated mutations; Host must match the configured origin host on every auth/resource request."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Authenticated"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "Host-only HttpOnly SameSite=Strict Path=/ cookie. Max-Age=28800 on login, 0 on logout. HTTPS cookie __Host-exhibitos_session has Secure; loopback HTTP cookie exhibitos_local_session does not.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            }
          },
          "400": {
            "description": "Malformed input, UUID or JSON schema; INVALID_INPUT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden resource/tenant, Origin/Host/CSRF rejection or rights denial.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds 1 MiB; BODY_TOO_LARGE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Generic internal failure; INTERNAL_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, revoked or unusable session; AUTH_REQUIRED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Login attempt or password-hash capacity limit; RATE_LIMITED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported request Content-Type; UNSUPPORTED_MEDIA_TYPE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "GetReadSession",
        "summary": "Read session",
        "description": "Returns current membership role and fixed eight-hour expiry.",
        "security": [
          {
            "localSession": []
          },
          {
            "networkSession": []
          }
        ],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Must equal AUTH_ORIGIN exactly when supplied. Required on login and authenticated mutations; Host must match the configured origin host on every auth/resource request."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            }
          },
          "400": {
            "description": "Malformed input, UUID or JSON schema; INVALID_INPUT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, revoked or unusable session; AUTH_REQUIRED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden resource/tenant, Origin/Host/CSRF rejection or rights denial.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Revision, subject, last-admin or multi-tenant-user conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds 1 MiB; BODY_TOO_LARGE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Generic internal failure; INTERNAL_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "PostLogoutSession",
        "summary": "Logout session",
        "description": "Revokes current session and clears cookie.",
        "security": [
          {
            "localSession": [],
            "csrfToken": []
          },
          {
            "networkSession": [],
            "csrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Must equal AUTH_ORIGIN exactly when supplied. Required on login and authenticated mutations; Host must match the configured origin host on every auth/resource request."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedOut"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "Host-only HttpOnly SameSite=Strict Path=/ cookie. Max-Age=28800 on login, 0 on logout. HTTPS cookie __Host-exhibitos_session has Secure; loopback HTTP cookie exhibitos_local_session does not.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            }
          },
          "400": {
            "description": "Malformed input, UUID or JSON schema; INVALID_INPUT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, revoked or unusable session; AUTH_REQUIRED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden resource/tenant, Origin/Host/CSRF rejection or rights denial.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Revision, subject, last-admin or multi-tenant-user conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds 1 MiB; BODY_TOO_LARGE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Generic internal failure; INTERNAL_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported request Content-Type; UNSUPPORTED_MEDIA_TYPE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "localSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "exhibitos_local_session",
        "description": "Local loopback HTTP only; HttpOnly SameSite=Strict Path=/."
      },
      "networkSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-exhibitos_session",
        "description": "HTTPS host-only Secure HttpOnly SameSite=Strict Path=/. Network listener remains loopback behind an external TLS terminator."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Synchronizer token obtained from authenticated session; required with session cookie for authenticated mutations."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "description": "Safe generic code message; internal database details are not returned."
          },
          "fieldErrors": {
            "type": "array",
            "maxItems": 0,
            "items": {}
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "fieldErrors",
          "requestId"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "subject": {
            "type": "string",
            "minLength": 3,
            "maxLength": 128,
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9_.@-]{2,127}$"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 1024,
            "writeOnly": true,
            "description": "At least 12 characters under request validation; UTF-8 bytes must also be between 12 and 1024. Never log or pass in an argv value."
          },
          "tenantId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          }
        },
        "required": [
          "subject",
          "password",
          "tenantId"
        ]
      },
      "Session": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "artist",
              "curator",
              "viewer"
            ]
          },
          "csrfToken": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$",
            "description": "Separate synchronizer token. Not a session bearer."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "tenantId",
          "role",
          "csrfToken",
          "expiresAt"
        ]
      },
...
      "Authenticated": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "const": true
          }
        },
        "required": [
          "authenticated"
        ]
      },
      "LoggedOut": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "loggedOut": {
            "const": true
          }
        },
        "required": [
          "loggedOut"
        ]
      },
...
    }
  }
}